Navigator: Add support for intercepting clicking links
When the user clicks an internal link in the Navigator the link might be to a resource not present in the reading order - example would be a note that was added with linear=no in the epub (Impossible to open a file when linear="no" #124).

This PR implements a way to intercept such a link click - and also for the listener to tell the Navigator not to navigate to the link - meaning that the app can implement handling the link itself (for example showing a popup with the contents of the link like iBooks do).

The implementation is done by extending the Navigator interface with a method called `shouldJumpToLink` that returns true by default and is only called when a link is clicked (when someone calls the `go` method with a link object - which is done when a link internal to the epub is clicked).